chore: code refactored and specs added
