+++ src/commands/deploy/internal/getDeployScripts.ts
@@ -1,35 +1,17 @@
-import { ServerType, Target } from '@sasjs/utils'
+import { Target } from '@sasjs/utils'
 import path from 'path'
 import { getConfiguration } from '../../../utils/config'
 
 export async function getDeployScripts(target: Target) {
-  const { buildSourceFolder, buildDestinationFolder } = process.sasjsConstants
+  const { buildSourceFolder } = process.sasjsConstants
   const configuration = await getConfiguration(
     path.join(buildSourceFolder, 'sasjs', 'sasjsconfig.json')
   )
 
-  let allDeployScripts: string[] = [
-    ...(configuration?.deployConfig?.deployScripts || [])
-  ]
-
-  allDeployScripts = [
-    ...allDeployScripts,
+  const allDeployScripts: string[] = [
+    ...(configuration?.deployConfig?.deployScripts || []),
     ...(target.deployConfig?.deployScripts || [])
   ]
 
-  if (
-    !!target.deployConfig?.deployServicePack &&
-    target.serverType === ServerType.Sas9
-  ) {
-    const buildOutputFileName =
-      target.buildConfig?.buildOutputFileName || `${target.name}.sas`
-    const deployScriptPath = path.join(
-      buildDestinationFolder,
-      buildOutputFileName
-    )
-    allDeployScripts = [...allDeployScripts, deployScriptPath]
-  }
-
-  allDeployScripts = allDeployScripts.filter((d) => !!d)
-  return [...new Set(allDeployScripts)]
+  return [...new Set(allDeployScripts.filter((d) => !!d))]
 }
