+++ src/commands/deploy/internal/index.ts
@@ -1,0 +1,8 @@
+export * from './deployToSASJSWithServicePack'
+export * from './getDeployScripts'
+export * from './executeDeployScript'
+export * from './executeSasScript'
+export * from './executeNonSasScript'
+export * from './deployToSasViya'
+export * from './deployToSas9'
+export * from './deployToSasjs'
